Extract browser-control JS builders/diagnostics into CmuxBrowserControl
Introduces the CmuxBrowserControl package and pulls the stateless half of the
TerminalController v2 browser RPC domain into a Sendable BrowserControlService,
matching the spec's "Service wraps the stateless JavaScript evaluation, element
finding, and diagnostics" shape.

What moved (pure, no god-type reach):
- jsonLiteral / normalizeJSValue (the latter takes an injected isUndefinedSentinel
  closure seam, since the sentinel type stays in the app target)
- failureLooksLikeCSPEvalBlock, describeJavaScriptError
- elementNotFoundMessage (the count-driven message branches)
- the not-found diagnostics JS probe builder
- the find.* harness (findScript) + the seven semantic finder bodies
  (role/text/label/placeholder/alt/title/testid) + findFirst/findLast/findNth
  selector scripts

Seams: TerminalController keeps all per-surface mutable state (element-ref table,
dialog queue, init scripts), the WebKit evaluation path, and the
Workspace/TabManager/BrowserPanel/WKWebView resolvers (section-6.3 stored-state
constraint). It owns a `nonisolated let v2BrowserControl` service and the moved
method bodies are now thin one-line forwards, so every v2 browser RPC reply is
byte-for-byte identical: the generated JS strings, the eval envelope keys
(__cmux_t/__cmux_v), the error/CSP classification, and the not-found message are
unchanged. The eval-envelope wire constants are injected from the existing
TerminalController statics via BrowserEvalEnvelope.

The larger stateful/WebKit-bound browser methods (navigate/click/screenshot/eval/
tabs/cookies/storage/dialogs/downloads) were intentionally left in place: they reach
TerminalController's private v2 infrastructure and app-only types and cannot move
without inverting ~15 seams or moving non-browser infra, which would not be
byte-identical-safe in one PR. This extracts the largest clean subset.

Package: swift-tools-version 6.0, v6 language mode, ExistentialAny +
InternalImportsByDefault; one major public type per file with full DocC; wired into
cmux.xcodeproj (6 pbxproj entries mirroring CmuxControlSocket/CmuxFoundation) and
added as an app-target dependency. Adds behavior unit tests (no fakes needed; the
logic is pure, plus a closure seam for the sentinel). TerminalController shrinks
287 lines; budget ratcheted. swift build + swift test green; package convention
lint OK with zero new lint:allow.

NOTE: full app build + tests validated by GitHub CI.
